Doesn't respond to bot anymore

+++ src/main/java/com/github/lrhel/rabbot/command/misc/EchoCommand.java
@@ -1,13 +1,15 @@
 package com.github.lrhel.rabbot.command.misc;
 
-import org.javacord.api.entity.message.Message;
-
 import de.btobastian.sdcf4j.Command;
 import de.btobastian.sdcf4j.CommandExecutor;
+import org.javacord.api.entity.message.Message;
+import org.javacord.api.entity.user.User;
 
 public class EchoCommand implements CommandExecutor {
 	@Command(aliases = {"echo"}, privateMessages = false, description = "Echo what I say!", showInHelpPage = false)
-	public String onEchoCommand(String[] arg, Message msg) {
+	public String onEchoCommand(User user, String[] arg, Message msg) {
+		if (user.isBot()) { return ""; }
+
 		if(msg.canYouDelete())
 			msg.delete();
